move getSources
